Updated models and migration files for investments and transactions
The updates were made in investments and transactions models and related migration files. This includes the addition of stock currency field in holding and domestic and foreign trade import functions in RakutenSecLoader class. Additionally, the update also includes the revised method for file retrieval and csv reading accommodating both development and production environments. Minor adjustments were also made in the templates and connectors.